backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
anthropic==0.18.1
python-dotenv==1.0.1
pydantic==2.5.3
